Initial commit
It all started with a simple mustache template many years ago. The
successor to mustache, "Handlebars" is a visual analog for a mustache
and successor to the format. "Two Whales Kissing" is another visual
analog for a mustache and since we're working with Ruby we could call
that, "Two Whales Kissing for Ruby", which is very long. Rhales
combines Ruby and Whales into a one-word name for our library. It's a
perfect name with absolutely no ambiguity or risk of confusion with
other gems.

Moved from onetimesecret/onetimesecret @ 167ec1da3bb4