Replace string concat with array
In parse_section_content along with a warning since it's parsing the
entire content of the section so should be replaced with StringScanner
like tokenize_content.

In parse_quoted_string, no such warning since the tag attribute
strings are so much shorter. It's probably more efficient to just
go character by character.